Проверка создания пользователя UI + Backend + DB

+++ src/main/kotlin/frontend/helpers/Extensions.kt
@@ -16,4 +16,4 @@ class Extensions {
     }
 }
 
-fun String.toPrice() = filter { it.isDigit() }.toFloat() / 100f
+fun String.toPrice() = filter { it.isDigit() }.toDouble() / 100
